python/otlp/requirements.txt
Deprecated==1.2.13
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
typing_extensions==4.4.0
wrapt==1.14.1
